fix(subscription): parse via serde_json, separate lib from main
